Adición de autenticación básica para acceder a los métodos

+++ infrastructure/receivers/entry-points/src/main/java/nisum/users/controller/UserController.java
@@ -2,6 +2,8 @@ package nisum.users.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.java.Log;
+import nisum.users.controller.auth.JWTGenerator;
+import nisum.users.controller.auth.JWTValidator;
 import nisum.users.controller.dto.ResponseDTO;
 import nisum.users.controller.dto.UserRequestDTO;
 import nisum.users.controller.dto.UserResponseDTO;
@@ -13,7 +15,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 @Log
@@ -23,6 +24,8 @@ import java.util.regex.Pattern;
 public class UserController {
 
     private final UserService userService;
+    private final JWTGenerator jwtGenerator;
+    private final JWTValidator jwtValidator;
 
     @Value("${regex.password}")
     private String passwordRegex;
@@ -31,8 +34,13 @@ public class UserController {
     private String emailRegex;
 
     @PostMapping("/save")
-    public ResponseEntity<ResponseDTO<UserResponseDTO>> saveUser(@RequestBody UserRequestDTO userDTO) {
+    public ResponseEntity<ResponseDTO<UserResponseDTO>> saveUser(
+            @RequestBody UserRequestDTO userDTO,
+            @RequestHeader("Authorization") String authorizationHeader) {
         try {
+            if (!jwtValidator.isValidToken(authorizationHeader))
+                return ResponseBuilder.build401Response();
+
             if (!isValidEmail(userDTO.getEmail()))
                 return ResponseBuilder.build400EmailResponse();
 
@@ -47,9 +55,14 @@ public class UserController {
     }
 
     @GetMapping("/user/{email}")
-    public ResponseEntity<ResponseDTO<UserResponseDTO>> findByEmail(@PathVariable("email") String email) {
+    public ResponseEntity<ResponseDTO<UserResponseDTO>> findByEmail(
+            @PathVariable("email") String email,
+            @RequestHeader("Authorization") String authorizationHeader) {
 
         try {
+            if (!jwtValidator.isValidToken(authorizationHeader))
+                return ResponseBuilder.build401Response();
+
             var userConsult = userService.findById(email);
             if (userConsult == null) return ResponseBuilder.build404Response();
             return ResponseBuilder.build200Response(UserMapper.mapUserToDTO(userConsult));
@@ -59,8 +72,12 @@ public class UserController {
     }
 
     @GetMapping("/view-users")
-    public ResponseEntity<ResponseDTO<List<UserResponseDTO>>> findAllUsers() {
+    public ResponseEntity<ResponseDTO<List<UserResponseDTO>>> findAllUsers(
+            @RequestHeader("Authorization") String authorizationHeader) {
         try {
+            if (!jwtValidator.isValidToken(authorizationHeader))
+                return ResponseBuilder.build401Response();
+
             var usersConsult = userService.findAll();
             var usersDtoConsult = usersConsult.stream().map(UserMapper::mapUserToDTO).toList();
             return ResponseBuilder.build200Response(usersDtoConsult);
@@ -70,8 +87,13 @@ public class UserController {
     }
 
     @PutMapping("/user/edit")
-    public ResponseEntity<ResponseDTO<UserResponseDTO>> updateUser(@RequestBody UserRequestDTO userDTO) {
+    public ResponseEntity<ResponseDTO<UserResponseDTO>> updateUser(
+            @RequestBody UserRequestDTO userDTO,
+            @RequestHeader("Authorization") String authorizationHeader) {
         try {
+            if (!jwtValidator.isValidToken(authorizationHeader))
+                return ResponseBuilder.build401Response();
+
             if (!isValidEmail(userDTO.getEmail()))
                 return ResponseBuilder.build400EmailResponse();
 
@@ -88,12 +110,29 @@ public class UserController {
     }
 
     @DeleteMapping("/user/delete/{email}")
-    public ResponseEntity<ResponseDTO<UserResponseDTO>> deleteByEmail(@PathVariable("email") String email) {
+    public ResponseEntity<ResponseDTO<UserResponseDTO>> deleteByEmail(
+            @PathVariable("email") String email,
+            @RequestHeader("Authorization") String authorizationHeader) {
 
         try {
+            if (!jwtValidator.isValidToken(authorizationHeader))
+                return ResponseBuilder.build401Response();
+
             var userConsult = userService.deleteById(email);
             if (userConsult == null) return ResponseBuilder.build404Response();
             return ResponseBuilder.build200DeletedResponse(UserMapper.mapUserToDTO(userConsult));
+        } catch (Exception ex) {
+            return ResponseBuilder.build500Response(ex.getMessage());
+        }
+    }
+
+    @GetMapping("/user/auth")
+    public ResponseEntity<ResponseDTO<String>> getJwtToken(
+            @RequestHeader("Secret") String secret) {
+        try {
+            var jwtGenerated = jwtGenerator.generate(secret);
+            if (jwtGenerated == null) return ResponseBuilder.build401Response();
+            return ResponseBuilder.build200Response(jwtGenerated);
         } catch (Exception ex) {
             return ResponseBuilder.build500Response(ex.getMessage());
         }
